Fixing up occurences of "unsigned long" in the main mps and test code so that we no longer assume that "unsigned long" is large enough for a pointer or size, or that it's the longest available integer type, since these are not true on 64-bit windows.
Copied from Perforce
 Change: 178021
 ServerID: perforce.ravenbrook.com

+++ mps/code/mpmtypes.h
@@ -132,7 +132,7 @@ typedef Res (*TraceFixMethod)(ScanState ss, Ref *refIO);
 
 /* This type is used by the PoolClass method Walk */
 typedef void (*FormattedObjectsStepMethod)(Addr obj, Format fmt, Pool pool,
-                                           void *v, unsigned long s);
+                                           void *v, size_t s);
 
 /* This type is used by the PoolClass method Walk */
 typedef void (*FreeBlockStepMethod)(Addr base, Addr limit, Pool pool, void *p);
